requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.1
frozenlist==1.6.0
greenlet==3.2.1
idna==3.10
lxml==5.4.0
multidict==6.4.3
propcache==0.3.1
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.7
SQLAlchemy==2.0.40
SQLAlchemy-Utils==0.41.2
typing_extensions==4.13.2
urllib3==2.4.0
xlrd==2.0.1
yarl==1.20.0
